+++ src/RequestParser/RequestParser.cpp
@@ -78,7 +78,7 @@ namespace AFG
     }
 
     /* splits client request in tokens with _delimiter. returns the token at position _token_pos. */
-    std::string RequestParser::parseToken(std::string _delimiter, int _token_pos)
+    std::string RequestParser::parseToken(std::string _delimiter, size_t _token_pos)
     {
         std::vector<std::string>    request_split_in_tokens;
 
@@ -94,7 +94,7 @@ namespace AFG
     1) gets split with space as the delimiter
     2) token specified through _token_pos (here "user1,user2") gets split with comma as the delimiter
     3) returns vector filled with the users (targets) */
-    std::vector<std::string> RequestParser::parseListToken(std::vector<std::string> _delimiter, int _token_pos)
+    std::vector<std::string> RequestParser::parseListToken(std::vector<std::string> _delimiter, size_t _token_pos)
     {
         std::vector<std::string>    request_split_in_tokens;
         std::vector<std::string>    token_list;
